// FE/CV_app_user/src/interfaces/IUser.ts
export interface IUser {
  id?: string;
  address?: string;
  avatar?: string;
  email?: string;
  username?: string;
  email_verified_at?: string;
  identity_verified?: boolean | null;
  information?: string;
  name?: string;
  social_id?: string;
  social_type?: string;
  telephone?: string;
  phone?: string;

  created_at?: string;
  updated_at?: string;
  created_by?: any;
  updated_by?: any;

  // addition
  isFollowed?: boolean;
  posts_count?: number;
}